+++ src/main/java/com/xwm/magicmaid/player/skill/perfomskill/secret/PerformSkillFlash.java
@@ -71,7 +71,7 @@ public class PerformSkillFlash extends PerformSkillSecretBase
     }
 
     @Override
-    public void drawIcon(int x, int y, float scale) {
+    public void drawIcon(float x, float y, float scale) {
         Minecraft.getMinecraft().getTextureManager().bindTexture(TEXTURE);
 
         double scalex = 46.0 / 46.0;
